# shared/openapi.yaml
openapi: 3.1.0
info:
  title: 뭐할구 API
  version: 1.0.0
  description: |
    서울 구별·카테고리별 추천 장소, 전체 장소 목록, 익명 게시판·댓글, 근거 기반 챗봇을 위한 공유 API 계약입니다.
    장소 랭킹은 시연용으로 선정한 추천 장소를 조합별 최대 5개 반환합니다.
servers:
  - url: http://localhost:8000
    description: 로컬 개발 서버
...
paths:
...
  /api/course-rankings:
    get:
      tags: [Courses]
      summary: 서울 고정 코스 랭킹 조회
      description: 강남구, 마포구, 서초구, 송파구, 노원구의 고정 코스 5개를 반환합니다.
      operationId: getCourseRankings
      responses:
        '200':
          description: 서울 코스 랭킹
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseRankingResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/course-suggestions:
    post:
      tags: [Courses]
      summary: 규칙 기반 코스 초안 생성
      description: 장소 좌표의 직선거리와 카테고리 균형을 이용하며 결과를 저장하지 않습니다.
      operationId: createCourseSuggestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseSuggestionRequest'
      responses:
        '200':
          description: 추천 코스 초안
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseSuggestionResponse'
        '400':
          description: 입력 오류 또는 추천 장소 부족
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/courses:
    post:
      tags: [Courses]
      summary: 익명 코스 저장
      operationId: createCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreateRequest'
      responses:
        '201':
          description: 저장된 코스
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/LocationNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/courses/{public_id}:
    parameters:
      - $ref: '#/components/parameters/CoursePublicIdPath'
    get:
      tags: [Courses]
      summary: 공유 코스 조회
      operationId: getCourse
      responses:
        '200':
          description: 저장된 코스
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/CourseNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Courses]
      summary: 공유 코스 수정
      operationId: updateCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdateRequest'
      responses:
        '200':
          description: 수정된 코스
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/PasswordMismatch'
        '404':
          description: 코스 또는 장소 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Courses]
      summary: 공유 코스 삭제
      operationId: deleteCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordRequest'
      responses:
        '204':
          description: 삭제 성공
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/PasswordMismatch'
        '404':
          $ref: '#/components/responses/CourseNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
...
    CoursePublicIdPath:
      name: public_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-f]{32}$'
        minLength: 32
        maxLength: 32
  responses:
    ValidationError:
      description: 입력값 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: 입력값을 확인해 주세요.
            details: null
    PasswordMismatch:
      description: 비밀번호 불일치
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: PASSWORD_MISMATCH
            message: 비밀번호가 일치하지 않습니다.
            details: null
...
    CourseNotFound:
      description: 코스 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: COURSE_NOT_FOUND
            message: 코스를 찾을 수 없습니다.
            details: null
    LocationNotFound:
      description: 장소 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: LOCATION_NOT_FOUND
            message: 장소를 찾을 수 없습니다.
            details: null
...
    InternalError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INTERNAL_ERROR
            message: 서버 오류가 발생했습니다.
            details: null
  schemas:
    LocationCategory:
      type: string
      enum: [관광지, 레포츠, 문화시설, 쇼핑, 숙박, 여행코스, 축제공연행사]
...
    LocationItem:
      type: object
      required: [content_id, category, title, address, district, longitude, latitude, image_url, thumbnail_url, phone]
      properties:
        content_id:
          type: string
        category:
          $ref: '#/components/schemas/LocationCategory'
        title:
          type: string
        address:
          type: [string, 'null']
        district:
          type: string
        longitude:
          type: [number, 'null']
        latitude:
          type: [number, 'null']
        image_url:
          type: [string, 'null']
          format: uri
        thumbnail_url:
          type: [string, 'null']
          format: uri
        phone:
          type: [string, 'null']
...
    CourseSuggestionRequest:
      type: object
      additionalProperties: false
      required: [district, categories, stop_count]
      properties:
        district:
          type: string
          minLength: 1
        categories:
          type: array
          minItems: 1
          maxItems: 3
          uniqueItems: true
          items:
            $ref: '#/components/schemas/LocationCategory'
        stop_count:
          type: integer
          minimum: 3
          maximum: 5
    CourseStop:
      type: object
      required: [position, distance_from_previous_meters, location]
      properties:
        position:
          type: integer
          minimum: 1
          maximum: 5
        distance_from_previous_meters:
          type: [integer, 'null']
          minimum: 0
        location:
          $ref: '#/components/schemas/LocationItem'
    CourseRankingItem:
      type: object
      required:
        - rank
        - district
        - title
        - description
        - thumbnail_url
        - stops
        - total_straight_line_distance_meters
      properties:
        rank:
          type: integer
          minimum: 1
          maximum: 5
        district:
          type: string
        title:
          type: string
        description:
          type: string
        thumbnail_url:
          type: [string, 'null']
          format: uri
        stops:
          type: array
          minItems: 3
          maxItems: 5
          items:
            $ref: '#/components/schemas/CourseStop'
        total_straight_line_distance_meters:
          type: [integer, 'null']
          minimum: 0
    CourseRankingResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 5
          maxItems: 5
          items:
            $ref: '#/components/schemas/CourseRankingItem'
    CourseSuggestionResponse:
      type: object
      required: [district, categories, stops, total_straight_line_distance_meters]
      properties:
        district:
          type: string
        categories:
          type: array
          minItems: 1
          maxItems: 3
          items:
            $ref: '#/components/schemas/LocationCategory'
        stops:
          type: array
          minItems: 3
          maxItems: 5
          items:
            $ref: '#/components/schemas/CourseStop'
        total_straight_line_distance_meters:
          type: integer
          minimum: 0
    CourseCreateRequest:
      type: object
      additionalProperties: false
      required: [title, password, location_content_ids]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
        password:
          type: string
          minLength: 4
          maxLength: 20
          writeOnly: true
        location_content_ids:
          type: array
          minItems: 3
          maxItems: 5
          uniqueItems: true
          items:
            type: string
    CourseUpdateRequest:
      type: object
      additionalProperties: false
      required: [title, password, location_content_ids]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
        password:
          type: string
          minLength: 4
          maxLength: 20
          writeOnly: true
        location_content_ids:
          type: array
          minItems: 3
          maxItems: 5
          uniqueItems: true
          items:
            type: string
    CourseDetail:
      type: object
      required: [public_id, title, created_at, updated_at, stops, total_straight_line_distance_meters]
      properties:
        public_id:
          type: string
          pattern: '^[0-9a-f]{32}$'
          minLength: 32
          maxLength: 32
        title:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        stops:
          type: array
          minItems: 3
          maxItems: 5
          items:
            $ref: '#/components/schemas/CourseStop'
        total_straight_line_distance_meters:
          type: [integer, 'null']
          minimum: 0
...
    PasswordRequest:
      type: object
      additionalProperties: false
      required: [password]
      properties:
        password:
          type: string
          minLength: 4
          maxLength: 20
          writeOnly: true
...
    ErrorResponse:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: [object, array, string, 'null']